PLD-5 #comment KB & Mouse events are processed fine, but gamepad events are wonky. Disconnecting it on the runtime freezes the window.

+++ Pleiad.Systems/IPleiadSystem.cs
@@ -6,6 +6,6 @@
         /// Function to execute on every update
         /// </summary>
         /// <param name="dTime">Time since the last update</param>
-        void Cycle(float dTime);
+        void Cycle(double dTime);
     }
 }
